gd32f103c8_drivers stm32_uart

+++ stm32f103c8_drivers/stm32f103c8_uart/driver/uart.h
@@ -2,10 +2,7 @@
 #define UART_DRV_H
 
 #include <stdint.h>
-#include <stdio.h>
 #include <stdbool.h>
-#include <stdarg.h>
-#include <string.h>
 
 #ifdef USE_STDPERIPH_DRIVER
 
@@ -47,30 +44,30 @@ typedef struct {
 
 /* 串口配置结构体 */
 typedef struct {
-	uart_periph_t uartx;    // 串口外设
-	uint32_t baud;			// 波特率
-	gpio_port_t tx_port;	// 发送端口
-	gpio_pin_t tx_pin;		// 发送引脚
-	gpio_port_t rx_port;	// 接收端口
-	gpio_pin_t rx_pin;		// 接收引脚
-    uint8_t *tx_buf;        // 发送数据缓冲区
-	uint8_t *rx_buf;        // 接收数据缓冲区
-	uint16_t tx_buf_size;   // 发送数据缓冲区大小
-	uint16_t rx_buf_size;   // 接收数据缓冲区大小
+	uart_periph_t uartx;
+	uint32_t baud;
+	gpio_port_t tx_port;
+	gpio_pin_t tx_pin;
+	gpio_port_t rx_port;
+	gpio_pin_t rx_pin;
+    uint8_t *tx_buf;
+	uint8_t *rx_buf;
+	uint16_t tx_buf_size;
+	uint16_t rx_buf_size;
 	uint16_t rx_single_max; // 单次接收最大数据量
 } uart_config_t;
 
 /* 串口设备结构体 */
 typedef struct uart_dev {
 	uart_config_t config;
-	bool init_flag;								// 初始化标志
-	uart_rx_cb_t rx_cb;							// 接收数据控制块
-	void (*printf)(char *format, ...);			// 格式化打印
-	void (*send)(uint8_t *data, uint32_t len);	// 发送数据
-	char *(*recv)(void);						// 接收数据
-	int8_t (*deinit)(struct uart_dev *dev);		// 去初始化
+	bool init_flag;
+	uart_rx_cb_t rx_cb;
+	void (*printf)(char *format, ...);
+	void (*send)(uint8_t *data, uint32_t len);
+	char *(*recv)(void);
+	int (*deinit)(struct uart_dev *dev);
 } uart_dev_t;
 
-int8_t uart_init(uart_dev_t *dev);
+int uart_init(uart_dev_t *dev);
 
 #endif
